sc drawstyles: Add a default style and make use of it
The style is empty, similar to the default cell style,
thus fall backs to the pool defaults for everything. This
ensures compatibility with existing documents, despite the
fact that the default style is applied to all shapes upon
import. In addition, people who will ignore styles and just
continue to use DF, will have their spreadsheets look the
same in older versions that don't support styles. For this
reason I also opted to not set a dedicated style for images
and OLE objects via
SdrModel::SetDefaultStyleSheetForSdrGrafObjAndSdrOle2Obj,
unlike Impress/Draw, and continue setting the "no fill and
no line" override as DF.

Change-Id: I11554044a1aaf386dc6c4acdbab798fc5a231adc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149359
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>