crashreporting: BitmapEx::operator=(BitmapEx const &) crashes
e.g.
https://crashreport.libreoffice.org/stats/crash_details/b68d5267-16b1-4ee9-bd42-bdcecdb7043c

this went wrong with:

commit bda56cd75a79bd4619e4fe60171c3c50eeb8429d
Date:   Sat Jun 25 10:21:17 2022 +1000

    vcl: extract function Animation::RenderNextFrameInAllRenderers()

    Change-Id: Iaccf12b251ffe06e24813c48ba3a169d0eaaa61c
    Reviewed-on: https://gerrit.libreoffice.org/c/core/+/76416

which substituted mnAnimCount with nAnimCount which is a subtle one char
difference, but mnAnimCount is a different thing entirely than the (then
local variable) nAnimCount. mnAnimCount is the count of animations
active, not the number of animation frames in a specific animation.

orig code was const size_t nAnimCount = maFrames.size();
so replace mnAnimCount here with maFrames.size() to get back to the
original logic

Change-Id: I11d1797eb0b9ef61addb3769d12d8364872071df
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149457
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>