tdf#115625 tdf#76022 sw table: CalcFlyOffset: get correct surround for textbox
The function thought it was dealing with a wrap-through shape,
but actually it is wrap parallel.

Attempts to syncProperty of RES_SURROUND meet with instant disaster,
and likely this emulation works better if the doubled-up shape
only has one wrap anyway.

There is a helper function for this after all,
so that highly suggests workarounds are necessary
instead of just syncing the wrap property.

There were two existing unit tests identified:
-tdf123163-1.docx (doesn't even open in MS Word 2010/2003)
-testCustomShapePresetExport.odt (on round-trip)
    -no differences noticed.

Change-Id: I7b0828ce6406fa9cfb1debb68040ea02e8d6fe45
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149463
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>