sw floattable: fix up/down cursor travel at fly boundary
Pressing the "down" key in the middle of the last line of the A1 cell on
page 1 went to the end of the paragraph instead of going to the A2 cell
on page 2.

The problem was that fly frames don't have a "next" pointer to the
follow fly frame, so the cursor travel code thought there is no follow
frame to jump to.

Fix the problem similar to how explicitly chained frames do this in
lcl_FindLayoutFrame().

While at it, also fix the the "up" direction.

Change-Id: I8dd0c2d011300520b126abf3ca8d17722ab5ff92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149476
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins