tdf#149485: Revert "tdf#142450 add code to store showing whitespace"
This reverts commit f48f0198a46d5948510a33a22b734aa21c6e594f.

The reverted commit claims that
"This option is stored at the document level like the other
layout options."
which is not true. The option is saved in the user profile,
which makes other documents to look wrong.
Reverting for now until it's implemented at document level
as the commit suggests

Change-Id: I2ae0daeb410090949f1b6cfc8245b711fe518216
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149541
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>