tdf#133089: Use proper Unicode code points for bullet style names
The names introduced in:

commit 5b1a4531d90bf9f3ed0c3629e415ab0fcc7d59e9
Date:   Tue Apr 2 08:30:17 2019 +0100

    tdf#124471 Change bullet styles names in Style section of Sidebar

Use Private Use Area code points from OpenSymbol font for the bullets,
but the code in sw/source/core/doc/DocumentStylePoolManager.cxx uses
proper non-PUA code points. PUA is problematic because they are
“private” as in Unicode does not define how they are interpret, so they
are font-specific and many systems will not do font fallback for them
because of this.

The names now use the non-PUA code points.

Change-Id: I462c054cff1ffbdfe66ad8d8172901b23fb4810c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149588
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>