tdf#150471 Remove duplicated Freeze commands from the Tabbed UI
As discussed in the ticket, the commands "Freeze First Column" and "Freeze First Row" are duplicated in the Tabbed UI (under the View tab), since there's already a "Freeze Rows and Columns" command that already offers these 2 options.

So this patch removes the "Freeze First Column" and "Freeze First Row" entries and keep the "Freeze Rows and Columns" comand.

The label of the ".uno:FreezePanes" is also changed to "Freeze Panes" so that it doesn't occupy so much space in the tab. I set the ContextLabel to "Freeze Rows and Columns", so in the menu and context menus, the text shown remains "Freeze Rows and Columns".

Change-Id: I57e6b22b103c2316ca0722878491b719e93b9bb4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/149650
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>