sw floattable: fix removal of last follow fly with a start/middle one
The bugdoc has 3 pages: each page has 1 row of a floating table.
Deleting the last row deleted the last fly frame, but the text in the
body didn't move from page 3 to page 2.

It seems what happened was that the condition in
SwTextFrame::HasNonLastSplitFlyDrawObj() was wrong, it always filtered
for the master anchor, not for the relevant anchor. Once that's fixed,
the other problem was that by the time SwRootFrame::DeleteEmptyFlys_()
would invalidate the anchor, the follow/precede relationship is already
unlinked for the to-be-deleted fly, so we can't find the correct anchor
of the to-be-deleted fly.

Fix the second problem by invalidating the anchor earlier, in
SwFlyAtContentFrame::DelEmpty(), where the to-be-deleted fly is still
part of the flow frame chain.

With this, deleting the last row in the document results in deleting the
entire page 3 and moving up the body text, as expected.

Change-Id: I3303408cafabcc61abfffe6c355ec3874fff30a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150008
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins