tdf#154565 ignore unwanted crossing event from search to treeview on "enter"
i.e if the crossing was triggered ultimately by a key stroke which is
likely from exiting the search box.

This way we can avoid the problem that with hover-selection that after
"enter" is used in the search box, selecting a matching row, then during
teardown of the widget the box is hidden, and the generated crossing
notification triggers unwanted selection of a different row under the
mouse.

If needs be this could be refined further to only happen for a specific
key or other details of the triggering event

Change-Id: I7e64fb5fa31a446b85016d86ecd314d13130d7b9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150016
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>