avmedia,*: guess the mime type of media files based on file name
.. at least for the most popular types, and do it automatically in
MediaItem::setURL().

This should work in practice in most cases and is much simpler than
adding some type detection or calling into platform dependent avmedia
backends.

Remove the parameter that was only ever set to
"application/vnd.sun.star.media" anyway, the same value that would be
used if it's missing.

Stop using that silly type for everything, only use it when guessing
fails.

In case an ODF document is loaded, it will use the mime type loaded from
the file (see setting of MediaMimeType in SdXMLPluginShapeContext) and
not guess it because that would require updating the entry in
manifest.xml as well.

Change-Id: I8ce29cf7425678ae11dda1d8c875be818f8623af
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150049
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>