sw: fix the other stale SwNumberPortion follow flag
This happens since commit f3a1bf473f2a7b9259223ba74f578b6cf0b1947d
on abi13630-1.abw.

itrpaint.cxx:419: void SwTextPainter::DrawTextLine: Assertion `roTaggedLabel' failed.

The problem is very similar to the situation fixed with commit
49ed57346b46dcd621f4470c32b8fff659777d04; then it was the pField which
ended up empty and deleted, while now it is "this" which ends up empty
and deleted; in either case the DrawTextLine checks of the follow flags
fail because some follow flag is erroneously set.

This requires a further tweak because the length of the portion is set
to 1, so Compress() returns the wrong result there but immediately after
SwFieldPortion::Format() returns SwNumberPortion::Format() sets the
length to 0 anyway, so simply set m_nFieldLen to 0 in ctor which is
correct anyway as there's no dummy char for these.

Change-Id: Ie99f67ee19a1f6019afe68764ec746cdc862984e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150055
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>