tdf#152234 drawinglayer: PDF/UA export: fallback for unsupported forms
veraPDF complains:

  Specification: ISO 14289-1:2014, Clause: 7.1, Test number: 3
  Content shall be marked as Artifact or tagged as real content

Form element can only be used for those form controls for which
PDFWriterImpl can produce a Widget annotation; see createDefaultWidget().

For a label control, it's not editable so it's not a PDF widget.

For an image button, it could be implemented in VCL but isn't currently.

So a fallback is required in any case, "Figure" or "Div" can be used.

Change-Id: I454030ff7c67a8a8101fcc59e945fe9dbc5f87a8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150059
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>