crashtesting: crash on saving of fdo48916-1.ods
That document has <style:style ... style:family="graphic">
before <style:default-style style:family="graphic">. The
problem was that only the default style was creating the
draw layer (in ScDrawDefaultsObj::getModelPool), which
means that the SdrItemPool didn't exist at the time the
item set of the former style was created.

The easiest solution seems to ensure the draw layer
existence on each UNO style object creation, similar to
the other draw layer related objects in this method.

A similar fix was added recently for another scenario in
commit 43ec0d511694304d7c610e6fcf196eb34e7c7b80 ("Ensure
valid SdrItemPool for drawing styles").

Change-Id: I50d89fea6d419119cd47ab163c82d1470b04b4a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150115
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>