tdf#154679 unable to scroll with mouse wheel with frozen rows
which is a problem since:

commit eedc3a207ec4b397ac2b3ed57887593dd880f9ad
Date:   Tue Mar 21 14:30:38 2023 +0000

    unmerge default and ScrollType::Drag cases for calc scrollbars

which noted that in

commit 11e5e3acfb3a530edf290200b2dc95d328b90be2
Date:   Thu Aug 4 16:38:47 2022 +0100

    nDelta overwritten on all branches except default

the default case was different to the Drag case

and speculated that the right approach might be to restore
that distinction.

but in fact ScrollBar::DoScroll(tools::Long nNewPos)
had an unconditional 'meScrollType = ScrollType::Drag' which
ScrollAdaptor::DoScroll doesn't so it was originally the case
that scroll wheels used to be reported to calc as "ScrollType::Drag"
so the default handler never arose and such scrolls were handled
by the drag handler after all, despite the unlikely looking scroll
type.

So restore the merge of ScrollType::Drag and default as the desired
outcome, dropping the comment about "only for warnings" which is
out of date now.

Change-Id: Ibb396f053ae518a6ad5a9b4266ee879c84883953
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150120
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>