uno:TextAttributes - add to popupmenu draw.xml for textboxes
Text Attributes includes margins for spacing the borders (lines)
away from the text.

I personally would like to see the order be Area, then lines,
then textAttributes, but I suppose one could look at it as
area margins instead of border margins.

In any case, margins are a pretty critical piece of either
lines or margins, and so should be a menu choice whenever
either of these is available.

Change-Id: Ifec5bbcd20dba6f42aba586bd0cd791bb9f3acae
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150131
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>