tdf#154703 writerfilter framePr: don't duplicate top/bot spacing
With framePr, all of the background/border settings
are really just paragraph properties.

BETTER would be to zero out the frame spacing
and leave it on the paragraph.
After all, there could be multiple paragraphs
and each could have their own spacing.

However, in this case we are getting the whole range as
one propertySet, so that basically means it is treated as one.
I presume.

But the biggest reason for zeroing out the paragraph instead of the frame
is because on export only the frame borders are being written out.
So that really needs to be fixed first.

make CppunitTest_sw_ooxmlexport10 CPPUNIT_TEST_NAME=testLibreOfficeHang

Change-Id: I1faeef6bf12668fb28e2b3710053ff545a4c36e8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150157
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>