tdf#154703 docx export framePr: export frame background
[This bug report is about exploring when Writer can export
in a way that imports as Writer frames instead of Draw objects.
Exporting the background is one of the most crucial pre-requisites.]

The framePr background always imports as fully transparent.
In that case it should just be ignored on export.

However, in the (theoretical) case of ODT->DOCX framePr,
or in the (practical) case where a user modiifies
a framePr's background colour, then that background should
be transferred to the non-backgrounded paragraphs in the frame.

make CppunitTest_sw_ooxmlexport18 CPPUNIT_TEST_NAME=testTdf154703_framePr2

Change-Id: I1f59267030810b56d2701822b7004069e02eb46d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150355
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>