qt: Don't show non-visible menu entries
Take the visibility of the menu item
(that e.g. gets set in `QtMenu::ShowItem`)
into account when updating the menu in
`QtMenu::DoFullMenuUpdate`.

With this in place, only the relevant entry is shown
in the context menu for items in the character table
in the special characters dialog (Writer: "Insert" -> "Special
Character") while previously, the "Add to favorites"
and "Remove from favorites" entries would both be shown
at the same time regardless of whether the character
was present in the favorites or not.
(Both entries would actually do the same: add the character
if it was not present in the favorites, remove it if it
was already there...).

Change-Id: Ica5f07d6009d97db2df0540293b096565bf1dd52
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150285
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>