sw floattable: limit the unfloat button to legacy ODF files
CppunitTest_sw_uiwriter2's testUnfloatButton and testUnfloating failed
in the SW_FORCE_FLY_SPLIT=1 case, on unfloatable_floating_table.doc.

The trouble was that it assumed that floating tables from DOC are
imported as tables in non-split flys, which is no longer the case.

Fix the problem by limiting the test to legacy ODT files.

Change-Id: I0c82c04f3e44d9d0c936a5eb386cd2c0e68f8af1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150395
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins